+++ test/worm_indicator_test.dart
@@ -3,13 +3,11 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:worm_indicator/worm_indicator.dart';
 
 void main() {
-  test('adds one to input values', () {
-    expect(() => WormIndicator(
-      length: 0,
-    ), throwsRangeError);
-    expect(() => WormIndicator(
-      length: 3,
-      controller: null,
-    ), throwsNullThrownError);
+  test('Constructor params', () {
+    expect(
+        () => WormIndicator(
+          length: 0,
+        ),
+        throwsAssertionError);
   });
 }